# spec/microservices/ms-02/04-api-contract.yaml
openapi: 3.1.0
info:
  title: Catalog and Product API
  version: 1.0.0
  description: |
    API contract for MS-02 Catalog and Product.
    This is the single source of truth for paths, fields, status codes, and response shapes.
servers:
  - url: http://localhost:8102/api/v1
...
paths:
  /products:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
    get:
      operationId: listProducts
      summary: List eligible or administrative products.
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/LanguageCode'
        - $ref: '#/components/parameters/CountryCode'
        - name: categoryId
          in: query
          schema: { type: string, format: uuid }
        - name: sku
          in: query
          schema: { type: string }
        - name: name
          in: query
          schema: { type: string }
        - name: manufacturerCode
          in: query
          schema: { type: string }
        - name: available
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Product page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createProduct
      summary: Create a product aggregate.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProductRequest' }
      responses:
        '201':
          description: Product created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /products/{productId}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/ProductId'
    get:
      operationId: getProduct
      summary: Get a product in a localized storefront context.
      security: []
      parameters:
        - $ref: '#/components/parameters/LanguageCode'
        - $ref: '#/components/parameters/CountryCode'
      responses:
        '200':
          description: Product.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    put:
      operationId: updateProduct
      summary: Update product metadata and associations.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProductRequest' }
      responses:
        '200':
          description: Updated product.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    delete:
      operationId: deleteProduct
      summary: Delete a product aggregate and dependents.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Deletion result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletionResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /products/slug/{friendlyUrl}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/FriendlyUrl'
    get:
      operationId: getProductByFriendlyUrl
      summary: Get a product by localized friendly URL.
      security: []
      parameters:
        - $ref: '#/components/parameters/LanguageCode'
        - $ref: '#/components/parameters/CountryCode'
      responses:
        '200':
          description: Product.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /products/sku/{sku}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - name: sku
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getProductBySku
      summary: Get a product by store-scoped SKU.
      security: []
      responses:
        '200':
          description: Product.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404': { $ref: '#/components/responses/NotFound' }
  /products/uniqueness:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
    get:
      operationId: checkProductSkuUniqueness
      summary: Check product SKU uniqueness.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: sku
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 200 }
        - name: excludeProductId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Existence result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExistsResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /products/{productId}/visibility:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/ProductId'
    patch:
      operationId: updateProductVisibility
      summary: Update visibility and purchase eligibility.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateVisibilityRequest' }
      responses:
        '200':
          description: Updated product.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /products/{productId}/categories/{categoryId}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/ProductId'
      - $ref: '#/components/parameters/CategoryId'
    post:
      operationId: attachProductToCategory
      summary: Attach a product to a category.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Updated product.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    delete:
      operationId: detachProductFromCategory
      summary: Detach a product from a category.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Updated product.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /categories/{categoryId}/products:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/CategoryId'
    get:
      operationId: listCategoryProducts
      summary: List products in a category subtree.
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/LanguageCode'
        - $ref: '#/components/parameters/CountryCode'
      responses:
        '200':
          description: Product page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductListResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /products/{productId}/variants:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/ProductId'
    get:
      operationId: listProductVariants
      summary: List variants.
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Variant page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductVariantListResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createProductVariant
      summary: Create a product variant.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateVariantRequest' }
      responses:
        '201':
          description: Variant created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductVariant' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /products/{productId}/variants/{variantId}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/ProductId'
      - $ref: '#/components/parameters/VariantId'
    get:
      operationId: getProductVariant
      summary: Get one product variant.
      security: []
      responses:
        '200':
          description: Variant.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductVariant' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateProductVariant
      summary: Update a product variant.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateVariantRequest' }
      responses:
        '200':
          description: Updated variant.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductVariant' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    delete:
      operationId: deleteProductVariant
      summary: Delete a product variant.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Deletion result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletionResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /products/{productId}/variants/uniqueness/{sku}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/ProductId'
      - name: sku
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: checkVariantSkuUniqueness
      summary: Check variant SKU uniqueness within a product.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Existence result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExistsResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /products/{productId}/options/price:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/ProductId'
    post:
      operationId: calculateProductOptionPrice
      summary: Calculate price for selected product options.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CalculatePriceRequest' }
      responses:
        '200':
          description: Calculated product price.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PriceResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
components:
  securitySchemes:
    bearerAuth:
      $ref: '../../shared/common-schemas.yaml#/components/securitySchemes/bearerAuth'
  parameters:
    TenantId:
      $ref: '../../shared/common-schemas.yaml#/components/parameters/TenantId'
    StoreId:
      $ref: '../../shared/common-schemas.yaml#/components/parameters/StoreId'
    CorrelationId:
      $ref: '../../shared/common-schemas.yaml#/components/parameters/CorrelationId'
...
    ProductId:
      name: productId
      in: path
      required: true
      schema: { type: string, format: uuid }
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema: { type: string, format: uuid }
    VariantId:
      name: variantId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    FriendlyUrl:
      name: friendlyUrl
      in: path
      required: true
      schema: { type: string, minLength: 1 }
    LanguageCode:
      name: languageCode
      in: query
      required: false
      schema: { type: string, default: en }
    CountryCode:
      name: countryCode
      in: query
      required: false
      schema: { type: string, minLength: 2, maxLength: 2 }
    Page:
      $ref: '../../shared/common-schemas.yaml#/components/parameters/Page'
    PageSize:
      $ref: '../../shared/common-schemas.yaml#/components/parameters/PageSize'
  responses:
    Unauthorized:
      description: Authentication required.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Insufficient catalog permission.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found or not eligible in the requested scope.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Business or concurrency conflict.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    UnprocessableEntity:
      description: Validation failure.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalError:
      description: Internal service failure.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ServiceUnavailable:
      description: Media or dependent provider unavailable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      $ref: '../../shared/common-schemas.yaml#/components/schemas/ErrorResponse'
    PaginationInfo:
      $ref: '../../shared/common-schemas.yaml#/components/schemas/PaginationInfo'
    ProductListResponse:
      allOf:
        - $ref: '../../shared/common-schemas.yaml#/components/schemas/ListEnvelope'
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Product' }
        pagination: { $ref: '#/components/schemas/PaginationInfo' }
...
    ProductVariantListResponse:
      allOf:
        - $ref: '../../shared/common-schemas.yaml#/components/schemas/ListEnvelope'
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ProductVariant' }
        pagination: { $ref: '#/components/schemas/PaginationInfo' }
    ExistsResponse:
      type: object
      required: [exists]
      properties:
        exists: { type: boolean }
    DeletionResult:
      type: object
      required: [id, status]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [Deleted] }
        dependentsRemoved: { type: integer, minimum: 0 }
        projectionEventPublished: { type: boolean }
...
    ProductDescription:
      type: object
      required: [languageCode, name, friendlyUrl]
      properties:
        languageCode: { type: string, minLength: 2 }
        name: { type: string, minLength: 1 }
        friendlyUrl: { type: string, minLength: 1 }
        description: { type: string }
        highlights: { type: string }
        title: { type: string }
        keywords: { type: string }
        metaDescription: { type: string }
    ProductMedia:
      type: object
      required: [id, fileName, imageType, mediaStatus]
      properties:
        id: { type: string, format: uuid }
        fileName: { type: string }
        imageType: { type: string, enum: [Binary, ExternalUrl] }
        originalUri: { type: string, format: uri }
        transformedUri: { type: string, format: uri }
        externalUrl: { type: string, format: uri }
        defaultImage: { type: boolean }
        mediaStatus: { type: string, enum: [Pending, Ready, Failed, Deleted] }
    Availability:
      type: object
      required: [id, regionCode, quantity, reservedQuantity, active]
      properties:
        id: { type: string, format: uuid }
        regionCode: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 0 }
        reservedQuantity: { type: integer, minimum: 0 }
        sellableQuantity: { type: integer, minimum: 0 }
        active: { type: boolean }
...
    Price:
      type: object
      required: [id, amount, currencyCode, defaultPrice]
      properties:
        id: { type: string, format: uuid }
        amount: { type: number, minimum: 0 }
        currencyCode: { type: string, minLength: 3, maxLength: 3 }
        priceType: { type: string, enum: [OneTime, Subscription, Rental] }
        defaultPrice: { type: boolean }
        specialAmount: { type: number, minimum: 0 }
        specialStartAt: { type: string, format: date-time }
        specialEndAt: { type: string, format: date-time }
        finalAmount: { type: number, minimum: 0 }
        discounted: { type: boolean }
    ProductOptionValue:
      type: object
      required: [id, code, displayOnly]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        displayOnly: { type: boolean }
        priceAdjustment: { type: number }
        imageUri: { type: string, format: uri }
    ProductOption:
      type: object
      required: [id, code, displayOnly, values]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        displayOnly: { type: boolean }
        values:
          type: array
          items: { $ref: '#/components/schemas/ProductOptionValue' }
    ProductVariant:
      type: object
      required: [id, productId, sku, status, defaultSelection]
      properties:
        id: { type: string, format: uuid }
        productId: { type: string, format: uuid }
        sku: { type: string }
        code: { type: string }
        status: { type: string, enum: [Draft, Active, Unavailable, Deleted] }
        defaultSelection: { type: boolean }
        available: { type: boolean }
        dateAvailable: { type: string, format: date-time }
        availability:
          type: array
          items: { $ref: '#/components/schemas/Availability' }
        price: { $ref: '#/components/schemas/Price' }
    Product:
      type: object
      required: [id, storeId, sku, status, visible, available, canBePurchased, dateAvailable, descriptions, availabilities]
      properties:
        id: { type: string, format: uuid }
        storeId: { type: string }
        sku: { type: string }
        refSku: { type: string }
        status: { type: string, enum: [Draft, Active, Unavailable, Deleted] }
        visible: { type: boolean }
        available: { type: boolean }
        canBePurchased: { type: boolean }
        dateAvailable: { type: string, format: date-time }
        manufacturerCode: { type: string }
        productTypeCode: { type: string }
        taxClassCode: { type: string }
        productVirtual: { type: boolean }
        productShippable: { type: boolean }
        productFree: { type: boolean }
        length: { type: number }
        width: { type: number }
        height: { type: number }
        weight: { type: number }
        reviewAverage: { type: number }
        reviewCount: { type: integer, minimum: 0 }
        sortOrder: { type: integer }
        descriptions:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ProductDescription' }
        categories:
          type: array
          items: { $ref: '#/components/schemas/CategoryReference' }
        variants:
          type: array
          items: { $ref: '#/components/schemas/ProductVariant' }
        availabilities:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Availability' }
        options:
          type: array
          items: { $ref: '#/components/schemas/ProductOption' }
        properties:
          type: array
          items: { $ref: '#/components/schemas/ProductProperty' }
        media:
          type: array
          items: { $ref: '#/components/schemas/ProductMedia' }
        price: { $ref: '#/components/schemas/Price' }
    ProductProperty:
      type: object
      required: [code, value]
      properties:
        code: { type: string }
        value: { type: string }
        name: { type: string }
    CategoryReference:
      type: object
      required: [id, code]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
...
    PriceSelection:
      type: object
      required: [optionId, valueId]
      properties:
        optionId: { type: string, format: uuid }
        valueId: { type: string, format: uuid }
    PriceResponse:
      type: object
      required: [finalAmount, currencyCode, priceSource, matchedSelections]
      properties:
        finalAmount: { type: number, minimum: 0 }
        originalAmount: { type: number, minimum: 0 }
        currencyCode: { type: string }
        priceSource: { type: string, enum: [Product, DefaultVariant, ParentProduct] }
        discounted: { type: boolean }
        matchedSelections: { type: integer, minimum: 0 }
    CalculatePriceRequest:
      type: object
      required: [selections]
      properties:
        selections:
          type: array
          items: { $ref: '#/components/schemas/PriceSelection' }
        currencyCode: { type: string }
        countryCode: { type: string }
    CreateProductRequest:
      type: object
      required: [sku, descriptions, availabilities]
      properties:
        sku: { type: string, minLength: 1, pattern: '^[A-Za-z0-9]+([_-][A-Za-z0-9]+)*$' }
        refSku: { type: string }
        visible: { type: boolean, default: false }
        canBePurchased: { type: boolean, default: true }
        dateAvailable: { type: string, format: date-time }
        manufacturerCode: { type: string }
        productTypeCode: { type: string }
        taxClassCode: { type: string }
        productVirtual: { type: boolean }
        productShippable: { type: boolean }
        productFree: { type: boolean }
        sortOrder: { type: integer }
        descriptions:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ProductDescription' }
        availabilities:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/AvailabilityInput' }
        categories:
          type: array
          items: { $ref: '#/components/schemas/CategoryReferenceInput' }
        variants:
          type: array
          items: { $ref: '#/components/schemas/CreateVariantRequest' }
    UpdateProductRequest:
      type: object
      minProperties: 1
      properties:
        sku: { type: string }
        refSku: { type: string }
        visible: { type: boolean }
        canBePurchased: { type: boolean }
        dateAvailable: { type: string, format: date-time }
        manufacturerCode: { type: string }
        productTypeCode: { type: string }
        taxClassCode: { type: string }
        productVirtual: { type: boolean }
        productShippable: { type: boolean }
        productFree: { type: boolean }
        sortOrder: { type: integer }
        descriptions:
          type: array
          items: { $ref: '#/components/schemas/ProductDescription' }
        availabilities:
          type: array
          items: { $ref: '#/components/schemas/AvailabilityInput' }
    UpdateVisibilityRequest:
      type: object
      required: [visible, canBePurchased]
      properties:
        visible: { type: boolean }
        canBePurchased: { type: boolean }
        dateAvailable: { type: string, format: date-time }
    AvailabilityInput:
      type: object
      required: [regionCode, quantity]
      properties:
        regionCode: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 0 }
        active: { type: boolean, default: true }
...
    CreateVariantRequest:
      type: object
      required: [sku]
      properties:
        sku: { type: string, minLength: 1 }
        code: { type: string }
        defaultSelection: { type: boolean, default: false }
        available: { type: boolean, default: false }
        dateAvailable: { type: string, format: date-time }
        availability:
          $ref: '#/components/schemas/AvailabilityInput'
    UpdateVariantRequest:
      type: object
      minProperties: 1
      properties:
        sku: { type: string }
        code: { type: string }
        defaultSelection: { type: boolean }
        available: { type: boolean }
        dateAvailable: { type: string, format: date-time }
...
    CategoryReferenceInput:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }